ClonnableBoards are working. Added a declareFunction primitive to declare a gobstones function. Added 'nroBolitas' and 'hayBolitas' gobstones functions.

+++ run-typescript.py
@@ -21,9 +21,11 @@ class CmdApplication(object):
 
     def initialize(self):
         self.input_key = _Getch()
+        self.options = None
+        self.arguments = None
         
     def usage(self):
-        return "No usage info."
+        return "No usage info.\n"
         
     def get_option_handlers(self):
         return {}
@@ -33,24 +35,26 @@ class CmdApplication(object):
         
     def error(self, message):
         sys.stderr.write(message)
-        
+            
+    def get_option_name(self, option_switch):
+        return remove_no(normalize_option_name(option_switch.split(" ")[0]))
+            
     def start(self):
         self.initialize()
-        arguments, options = get_options(self.get_option_switches())
+        self.arguments, self.options = get_options(self.get_option_switches())
         
-        option_processed = None
-        
+        options_processed = []
         for option_switch, option_handler in self.get_option_handlers().iteritems():
-            option = option_switch.split(" ")[0]
-            if options[option]:
-                if isinstance(options[option], list):
-                    option_handler(*options[option])
+            option = self.get_option_name(option_switch)
+            if self.options[option] and not option_handler is None:
+                if isinstance(self.options[option], list):
+                    option_handler(*self.options[option])
                 else:
                     option_handler()
-                option_processed = option
-                break
+        
+                options_processed += [option]
             
-        if option_processed is None:
+        if len(options_processed) == 0:
             self.error(self.usage().replace('<CMDLINE>', sys.argv[0]))
             sys.exit(1)
 
@@ -68,6 +72,18 @@ def delete_file(filename):
     os.unlink(filename)
 
 ## Parser for option switches
+
+def remove_no(option_name):
+    if option_name[:3] == "no-":
+        return option_name[3:]
+    else:
+        return option_name
+
+def normalize_option_name(option_name):
+    if option_name[:2] == "--":
+        return option_name[2:]
+    else:
+        return option_name
 
 def get_options(option_switches):
     return parse_options(option_switches, sys.argv)
@@ -88,19 +104,25 @@ def default_options(option_switches):
             opt[sw] = []
     return opt
 
+def get_option_names(option_switches):
+    return [option.split(' ')[0] for option in option_switches]
+
 def parse_options(option_switches, args, max_args=None):    
     arguments = []
-    opt = default_options(option_switches)
+    option_names = get_option_names(option_switches)
+    switches = [normalize_option_name(s) for s in option_switches]
+    args = map(lambda arg: normalize_option_name(arg) if arg in option_names else arg, args)
+    opt = default_options(switches)    
     i = 1
     n = len(args)
-    
+        
     # all args
     while i < len(args):
         o = None        
         
         # select matching option
-        for oi in option_switches:
-            oi = oi.split(' ')
+        for oi in switches:
+            oi = oi.split(' ')            
             if oi[0] == args[i]:
                 o = oi
                 break
@@ -115,11 +137,11 @@ def parse_options(option_switches, args, max_args=None):
 
         # Check if single-word option
         # sw = current option
-        sw = o[0]
+        sw = o[0]                
         if len(o) == 1:
-            if sw[:3] == 'no-':
+            if sw[:3] == 'no-': 
                 neg = True
-                sw = sw[3:]
+                sw = sw[3:]            
             else:
                 neg = False
             opt[sw] = not neg
@@ -189,7 +211,9 @@ class RunTypescript(CmdApplication):
     def get_option_handlers(self):
         return {
             "--main X" : self.build,
-            "--purge"  : self.purge
+            "--purge"  : self.purge,
+            "--no-build-html" : None,
+            "--no-clear" : None
         }
         
     def purge(self):
@@ -204,21 +228,25 @@ class RunTypescript(CmdApplication):
         os.system("tsc %s" % (mainfile,))
         
         bin_dir = os.path.join(os.getcwd(), "bin")
-        if os.path.exists(bin_dir):
-            shutil.rmtree(bin_dir)    
-        os.mkdir(bin_dir)
+
+        if self.options["clear"]:
+            if os.path.exists(bin_dir):
+                shutil.rmtree(bin_dir)    
+            os.mkdir(bin_dir)
+            
         jsfiles = get_files_matching(os.path.dirname(mainfile), "*.js")        
         jsfiles = filter(lambda f: not mainfile[:-3] + ".js" in f, jsfiles)
         for f in jsfiles:
             os.system("mv %s %s" % (f, bin_dir))        
         os.system("mv %s %s" % (mainfile[:-3]+".js", bin_dir))
         
-        jsfiles = map(lambda f: os.path.basename(f), jsfiles)
-                    
-        html = self.render_html(os.path.basename(mainfile)[:-3]+".js", jsfiles)
-        write_file(os.path.join(bin_dir, "index.html"), html)
-        #os.system("google-chrome %s/index.html" % (os.getcwd(),))
-        #os.system("cat %s/index.html" % (os.getcwd(),))
+        if self.options["build-html"]:
+            print "Building html file..."
+            jsfiles = map(lambda f: os.path.basename(f), jsfiles)                    
+            html = self.render_html(os.path.basename(mainfile)[:-3]+".js", jsfiles)
+            write_file(os.path.join(bin_dir, "index.html"), html)
+            #os.system("google-chrome %s/index.html" % (os.getcwd(),))
+            #os.system("cat %s/index.html" % (os.getcwd(),))
         print "Ready"
 
     def render_html(self, mainf, jsfiles):
